Renamed from UserPreferences and add Injection.
